Function to display a bland plot in order to visually assess the agreement between CytOpt estimation of the class proportions and the estimate of the class proportions provided through manual gating.

Function to display a bland plot in order to visually assess the agreement between CytOpt estimation of the class proportions and the estimate of the class proportions provided through manual gating.

Usage

barplot_prop(proportions, title = "", xaxis_angle = 45)

Arguments

proportions

data.frame of (true and) estimated proportions from CytOpt()

title

plot title. Default is "", i.e. no title.

xaxis_angle

scalar indicating an angle to tilt the labels of x_axis. Default is 45.

Value

a ggplot object

Examples

if(interactive()){

res <- CytOpT(X_s = HIPC_Stanford_1228_1A, X_t = HIPC_Stanford_1369_1A, 
             Lab_source = HIPC_Stanford_1228_1A_labels,
             eps = 0.0001, lbd = 0.0001, n_iter = 10000, n_stoc=10,
             step_grad = 10, step = 5, power = 0.99, 
             method='minmax')
barplot_prop(res$proportions)

}