Function to estimate the type cell proportions in an unclassified cytometry data set denoted X_s by using the classification Lab_source from an other cytometry data set X_s. With this function an additional regularization parameter on the class proportions enables a faster computation of the estimator.

Function to estimate the type cell proportions in an unclassified cytometry data set denoted X_s by using the classification Lab_source from an other cytometry data set X_s. With this function an additional regularization parameter on the class proportions enables a faster computation of the estimator.

Usage

cytopt_minmax_r(
  X_s,
  X_t,
  Lab_source,
  theta_true = NULL,
  eps = 1e-04,
  lbd = 1e-04,
  n_iter = 10000,
  step = 5,
  power = 0.99,
  monitoring = FALSE
)

Arguments

X_s

Cytometry data set. The columns correspond to the different biological markers tracked. One line corresponds to the cytometry measurements performed on one cell. The classification of this Cytometry data set must be provided with the Lab_source parameters.

X_t

Cytometry data set. The columns correspond to the different biological markers tracked. One line corresponds to the cytometry measurements performed on one cell. The CytOpT algorithm targets the cell type proportion in this Cytometry data set.

Lab_source

Classification of the X_s Cytometry data set

theta_true

If available, gold-standard proportions in the target data set X_t derived from manual gating. It allows to assess the gap between the estimate and the gold-standard. Default is NULL, in which case no assessment is performed.

eps

Regularization parameter of the Wasserstein distance

lbd

an float constant that multiply the step-size policy. Default is 1e-04.

n_iter

an integer Constant that iterate method select. Default is 10000.

step

Constant that multiply the step-size policy. Default is 5.

power

the step size policy of the gradient ascent method is step/n^power. Default is 0.99.

monitoring

boolean indicating whether Kullback-Leibler divergence should be monitored and store throughout the optimization iterations. Default is FALSE.

Value

A list with the following elements:Results_Minmax